% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if a cache path exists}
\usage{
cache_exists(path)
}
\arguments{
\item{path}{File path.}
}
\value{
Logical indicating existence.
}
\description{
Check if a cache path exists
}
