% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_events}
\alias{plot_events}
\title{Plot standardized events}
\usage{
plot_events(events, league, ...)
}
\arguments{
\item{events}{Events data frame with x_unit/y_unit.}

\item{league}{League identifier.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot standardized events
}
