% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_field}
\alias{plot_field}
\title{Plot helpers
Plot a league field/court outline}
\usage{
plot_field(league, ...)
}
\arguments{
\item{league}{League identifier.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Plot helpers
Plot a league field/court outline
}
