% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-schemas.R
\name{schemas}
\alias{schemas}
\title{Get schema paths for a league}
\usage{
schemas(league)
}
\arguments{
\item{league}{League identifier (nba/nhl/nfl/mlb).}
}
\value{
Named list of schema paths.
}
\description{
Get schema paths for a league
}
