% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-nba-espn.R
\name{source_nba_espn}
\alias{source_nba_espn}
\title{NBA ESPN source helpers (internal)}
\usage{
source_nba_espn()
}
\description{
Internal ESPN NBA source implementation returned by \code{get_source("nba", source = "espn")}.
}
\keyword{internal}
