% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-nba-stats.R
\name{source_nba_stats}
\alias{source_nba_stats}
\title{NBA Stats source helpers (internal)}
\usage{
source_nba_stats()
}
\description{
NBA Stats source implementation returned by \code{get_source("nba", source = "nba_stats")}.
}
\keyword{internal}
