% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-nhl.R
\name{source_nhl}
\alias{source_nhl}
\title{NHL source helpers (internal)}
\usage{
source_nhl()
}
\description{
Internal ESPN NHL source implementation returned by \code{get_source("nhl")}.
}
\keyword{internal}
